refactor(types): seperate specific backend types
