+++ packages/lib/src/types/blaze.ts
@@ -1,0 +1,8 @@
+export type BlazeOption = {
+    name: string;
+    url: string;
+    auth?: string;
+    backendMeasures: string;
+    uiSiteMap: string[][];
+    catalogueKeyToResponseKeyMap: string[][];
+};
